# app.js
// Import the centralized configuration
const config = require('./config');

const express = require('express');
const path = require('path');
const mongoose = require('mongoose');
const ejsMate = require('ejs-mate');
const methodOverride = require('method-override'); // to use PUT and DELETE requests in forms
// JWT authentication is now used instead of passport sessions
const mongoSanitize = require('express-mongo-sanitize'); // sanitize user input to prevent NoSQL Injection
const helmet = require('helmet'); // to set various HTTP headers for security (security middleware)
const cors = require('cors'); // to allow cross-origin requests
const cookieParser = require('cookie-parser'); // to parse cookies
const { setFlashMessages, readFlashMessages } = require('./middleware/flashMessages'); // custom flash messages middleware
const swaggerUi = require('swagger-ui-express'); // for API documentation
const swaggerDocument = require('./docs/swagger.json'); // swagger documentation

const campgroundRoutes = require('./routes/campgrounds');
const reviewRoutes = require('./routes/reviews');
const userRoutes = require('./routes/users');
const bookingRoutes = require('./routes/bookings');
const User = require('./models/user');
const adminRoutes = require('./routes/admin');
const { addBookingCountToUser, addConfigToTemplates } = require('./middleware');

const ExpressError = require('./utils/ExpressError'); // import the ExpressError class from the utils folder
const ApiResponse = require('./utils/ApiResponse'); // import the ApiResponse class for standardized API responses

// Connect to MongoDB using configuration
// Using async IIFE for better error handling with async/await
(async () => {
  try {
    await mongoose.connect(config.db.url);
    console.log('Database connected to:', config.db.url);
  } catch (err) {
    console.error('MongoDB Atlas connection error:', err.message);

    // In development, try to connect to local MongoDB if Atlas connection fails
    if (!config.server.isProduction) {
      try {
        const localMongoUrl = 'mongodb://localhost:27017/myan-camp';
        console.log('Attempting to connect to local MongoDB...');
        await mongoose.connect(localMongoUrl);
        console.log('Connected to local MongoDB:', localMongoUrl);
      } catch (localErr) {
        console.error('Local MongoDB connection also failed:', localErr.message);
        console.error(
          'Please ensure MongoDB is running locally or check your Atlas connection string'
        );
      }
    } else {
      // In production, exit the application if database connection fails
      console.error('Exiting application due to database connection failure');
      process.exit(1);
    }
  }
})();

const app = express(); // create an instance of express app to use its methods

app.engine('ejs', ejsMate); // set ejs-mate as the engine for ejs files
app.set('view engine', 'ejs');
app.set('views', path.join(__dirname, 'views')); // set the views directory

app.use(express.urlencoded({ extended: true })); // to parse the form data
app.use(express.json()); // to parse JSON data
app.use(methodOverride('_method'));
app.use(express.static(path.join(__dirname, 'public'))); // to serve static files like CSS, JS, images
app.use(mongoSanitize({ replaceWith: '_' })); // sanitize user input to prevent NoSQL Injection

// Configure CORS using centralized config
app.use(cors(config.cors));

app.use(cookieParser()); // to parse cookies
app.use(setFlashMessages); // to set flash messages in cookies
app.use(helmet()); // to set various HTTP headers for security

// Content Security Policy (CSP) to prevent XSS attacks using centralized config
app.use(
  helmet.contentSecurityPolicy({
    directives: {
      defaultSrc: [],
      connectSrc: ["'self'", ...config.csp.connectSrc],
      scriptSrc: ["'unsafe-inline'", "'self'", ...config.csp.scriptSrc],
      styleSrc: ["'self'", "'unsafe-inline'", ...config.csp.styleSrc],
      workerSrc: ["'self'", 'blob:'],
      objectSrc: [],
      imgSrc: ["'self'", ...config.csp.imgSrc],
      fontSrc: ["'self'", ...config.csp.fontSrc],
      formAction: ["'self'", ...config.csp.formAction],
      reportUri: '/csp-violation-report-endpoint',
      frameSrc: ["'self'"],
      upgradeInsecureRequests: [],
      baseUri: ["'self'"],
    },
  })
);

app.post('/csp-violation-report-endpoint', express.json(), (req, res) => {
  console.log('CSP Violation:', req.body);
  res.status(204).end();
});

// JWT authentication is now used instead of passport sessions
app.use(addBookingCountToUser);
app.use(addConfigToTemplates); // Add configuration to res.locals for use in templates

// Make current user available in all templates
app.use((req, res, next) => {
  res.locals.currentUser = req.user;
  next();
});

// Read flash messages from cookies and make them available in templates
app.use(readFlashMessages);

// Import API versioning middleware
const { versionRoutes, deprecateEndpoint } = require('./middleware/apiVersioning');
const { authenticateJWT } = require('./middleware/jwtAuth');
const { apiLimiter } = require('./middleware/rateLimiter');

// Apply rate limiting to all API routes
app.use('/api', apiLimiter);

// Apply JWT authentication middleware to all API routes
// This middleware doesn't block requests without a JWT token, it just sets req.user if a token is provided
console.log('Applying JWT authentication middleware to all API routes');
app.use('/api', authenticateJWT);

// Apply versioning middleware to all API routes
app.use('/api', versionRoutes());

// API Routes
const apiV1Routes = require('./routes/api/v1');

// Traditional Routes
app.use('/campgrounds/:id/reviews', reviewRoutes); // use the review routes
app.use('/campgrounds', campgroundRoutes); // use the campground routes
app.use('/bookings', bookingRoutes);
app.use('/admin', adminRoutes);
app.use('/', userRoutes); // use the user routes

// Versioned API Routes
app.use('/api/v1', apiV1Routes); // use the versioned API routes

// API Documentation
app.use(
  '/api/docs',
  swaggerUi.serve,
  swaggerUi.setup(swaggerDocument, {
    explorer: true,
    customCss: '.swagger-ui .topbar { display: none }',
    swaggerOptions: {
      docExpansion: 'none',
      persistAuthorization: true,
    },
  })
);

app.get('/', (req, res) => {
  res.render('home');
});

// Serve React app
if (process.env.NODE_ENV === 'production') {
  app.use(express.static(path.join(__dirname, 'public/dist')));

  // For any routes that don't match API or traditional routes, serve the React app
  app.get('/app/*', (req, res) => {
    res.sendFile(path.join(__dirname, 'public/dist/index.html'));
  });
} else {
  // In development mode, handle React app routes
  app.get('/app/*', (req, res) => {
    res.redirect('http://localhost:5173' + req.originalUrl);
  });
}

app.all('*', (req, res, next) => {
  // catch all route
  if (req.originalUrl.startsWith('/api')) {
    return ApiResponse.error(
      'API endpoint not found',
      'The requested API endpoint does not exist',
      404
    ).send(res);
  }
  next(new ExpressError('Page Not Found', 404)); // pass the error to the error handler middleware
});

// Import the error handler middleware
const { errorHandler } = require('./utils/errorHandler');

// Use the centralized error handler middleware
app.use(errorHandler);

// Function to start the server using port from centralized config
const startServer = (port) => {
  return new Promise((resolve, reject) => {
    const server = app
      .listen(port, () => {
        console.log(`Serving on Port ${port}...`);
        resolve(server);
      })
      .on('error', (err) => {
        if (err.code === 'EADDRINUSE') {
          console.log(`Port ${port} is already in use, trying another port...`);
          reject(err);
        } else {
          reject(err);
        }
      });
  });
};

// Try to start the server on the specified port, or try alternative ports
const tryStartServer = async (initialPort, maxAttempts = 5) => {
  let currentPort = initialPort;
  let attempts = 0;

  while (attempts < maxAttempts) {
    try {
      const server = await startServer(currentPort);
      return server;
    } catch (err) {
      if (err.code === 'EADDRINUSE') {
        attempts++;
        currentPort++;
      } else {
        console.error('Failed to start server:', err);
        process.exit(1);
      }
    }
  }

  console.error(`Could not find an available port after ${maxAttempts} attempts.`);
  process.exit(1);
};

// Start the server with port from config
tryStartServer(config.server.port);