# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MyanCamp API",
    "description": "API documentation for MyanCamp - a campground booking platform",
    "version": "1.0.0",
    "contact": {
      "name": "MyanCamp Support",
      "email": "support@myancamp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://myancamp.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/admin/bookings": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all bookings",
        "description": "Returns a paginated list of all bookings",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by booking status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["confirmed", "pending", "cancelled", "completed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bookings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Booking"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 5
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 45
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Bookings retrieved successfully"
                    },
                    "error": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/campgrounds": {
      "get": {
        "tags": ["Campgrounds"],
        "summary": "Get all campgrounds",
        "description": "Returns a list of all campgrounds and distinct locations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "campgrounds": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Campground"
                          }
                        },
                        "locations": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Campgrounds retrieved successfully"
                    },
                    "error": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campgrounds"],
        "summary": "Create a new campground",
        "description": "Creates a new campground with the provided data",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "campground": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "Inle Lake Campground"
                      },
                      "location": {
                        "type": "string",
                        "example": "Inle Lake, Shan State, Myanmar"
                      },
                      "description": {
                        "type": "string",
                        "example": "A beautiful campground by the lake"
                      },
                      "price": {
                        "type": "number",
                        "example": 20
                      }
                    },
                    "required": ["title", "location", "description", "price"]
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "campground[title]": {
                    "type": "string",
                    "example": "Inle Lake Campground"
                  },
                  "campground[location]": {
                    "type": "string",
                    "example": "Inle Lake, Shan State, Myanmar"
                  },
                  "campground[description]": {
                    "type": "string",
                    "example": "A beautiful campground by the lake"
                  },
                  "campground[price]": {
                    "type": "number",
                    "example": 20
                  },
                  "image": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campground created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "campground": {
                          "$ref": "#/components/schemas/Campground"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Campground created successfully"
                    },
                    "error": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Booking": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c85"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2023-07-15"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2023-07-20"
          },
          "totalPrice": {
            "type": "number",
            "example": 250
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled", "completed"],
            "example": "confirmed"
          },
          "user": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c85"
              },
              "username": {
                "type": "string",
                "example": "johndoe"
              },
              "email": {
                "type": "string",
                "example": "john@example.com"
              }
            }
          },
          "campground": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c85"
              },
              "title": {
                "type": "string",
                "example": "Inle Lake Campground"
              },
              "location": {
                "type": "string",
                "example": "Inle Lake, Shan State, Myanmar"
              }
            }
          },
          "campsite": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c85"
              },
              "name": {
                "type": "string",
                "example": "Lakeside Tent Spot #3"
              },
              "price": {
                "type": "number",
                "example": 50
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T10:15:18.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-06-01T10:15:18.000Z"
          }
        }
      },
...
      "Campground": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c85"
          },
          "title": {
            "type": "string",
            "example": "Inle Lake Campground"
          },
          "price": {
            "type": "number",
            "example": 20
          },
          "description": {
            "type": "string",
            "example": "A beautiful campground by the lake"
          },
          "location": {
            "type": "string",
            "example": "Inle Lake, Shan State, Myanmar"
          },
          "geometry": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "Point"
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "example": [96.9102, 20.5862]
              }
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "example": "https://res.cloudinary.com/dlvtzyb7j/image/upload/v1624456518/MyanCamp/campground1_xyzabc.jpg"
                },
                "filename": {
                  "type": "string",
                  "example": "MyanCamp/campground1_xyzabc"
                }
              }
            }
          },
          "author": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c85"
              },
              "username": {
                "type": "string",
                "example": "admin"
              }
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "60d21b4667d0d8992e610c85"
                },
                "body": {
                  "type": "string",
                  "example": "Great place to stay!"
                },
                "rating": {
                  "type": "number",
                  "example": 5
                },
                "author": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "60d21b4667d0d8992e610c85"
                    },
                    "username": {
                      "type": "string",
                      "example": "user123"
                    }
                  }
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-06-23T10:15:18.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-06-23T10:15:18.000Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string",
            "example": "An error occurred while processing your request"
          },
          "error": {
            "type": "string",
            "example": "Campground not found"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}